refactor: add a /results sub path to avoid rerenders
