--- src/app/demo/search/@search/default.tsx
@@ -1,3 +1,0 @@
-export default function DefaultSearch() {
-  return null;
-}
