+++ src/app/demo/search/default.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function DefaultSearch() {
+  return redirect("/demo/search");
+}
